turn off all cloud sync features by default
with the first time setup screen, it feels more reasonable to wait for
the user to opt-in to having this feature enabled than it did with this
only being in a somewhat tucked away menu.

it would be nice to fetch player data from the cloud if they don't have
any data locally when opting in, but that's something that could
probably be done at a later point.